initial work on story 1313, ClientService and ClientAttendanceDao
also some work on preparing to upgrade to JUnit 4.4 by disambiguating some Integer Asserts.



git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@14408 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ mifos/src/test/java/org/mifos/application/accounts/business/AddAccountActionTest.java
@@ -4,6 +4,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.mifos.framework.util.helpers.TestObjectFactory.TEST_LOCALE;
+import junit.framework.Assert;
 import junit.framework.JUnit4TestAdapter;
 
 import org.hibernate.Session;
@@ -38,7 +39,7 @@ public class AddAccountActionTest {
 				AccountActionEntity.class, SEND_TO_ORPHANS);
 		action.setLocaleId(TEST_LOCALE);
 
-		assertEquals(SEND_TO_ORPHANS, action.getId());
+		assertEquals(SEND_TO_ORPHANS, (short) action.getId());
 		assertEquals(" ", action.getLookUpValue().getLookUpName());
 	}
